docker/book_rec_model/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
Cython==0.29.33
dvc[s3]==2.56.0
fastapi==0.96.0
idna==3.4
joblib==1.2.0
lightfm==1.16
mlflow==2.3.2
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
urllib3==1.26.15
uvicorn==0.22.0
